fix bug affecting uninitialized variables in functions that get called more than once test case that shows this bug: test/programs/simple/uninitializedVariableUsed3_BUG.cil.c
WARNING:
This fix greatly affects the performance of our standard test cases (the simplified SSH ones). One example (s3_srvr1_1.cil.c) doesn't terminate with this fix although the change shouldn't make a difference in this case. We don't know why this happens, but the change is definitely needed to fix the above bug, so we commit it.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@1709 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c