table-generator modified: link to logfile added.
now in the html-table the first column of every test is the logfile-column, which contains a link to the logfile of the test.
there is no check for an existing logfile in the script. if the logfile does not exist, the browser says "file not found", if you click the link.

as the logfile is only referenced (a typical html-link), it can be newer than the table and be inconsistent with the table.
if you first run a benchmark, create a table and then run the benchmark again without recreating the table,
the old logfiles are overwritten and the links in the old table will reference the new logfiles.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@3995 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c