table-generator modified: middleLine is movable now.
the middleLine between the two columns can be moved sidewards like a window border. 
just move the small black line between the columnHeaders, 
it will become green (hover) and orange (moving). 
colors are for easier testing and can be changed later :-)

however there is a little bug, because of the use of iframes in the art- and cfa-tab: 
if you press the mouse on the middleLine and move the cursor into the iframe, you can release the mouse, and the middleLine follows your cursor if you leave the iframe again. than you have to click outside the iframe again to stop this behaviour.
the reason for this is, that iframes (and their content) are not part of the current document.

because of this bug, the middleLine is only between the headers (toolbars) and not between the bodies (errorpathwin, tabs) of the columns.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@4027 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c