report-template modified: zoomFunctions changed.
Before this commit, zooming was done by scaling the SVG with an attribute 'scale'. Now zooming is done by calculating new height and width. This changes the size of the iframe, too, and there is no free place under or right of the SVG when zooming in higher levels. 
The viewport should be fixed (left upper corner before zooming is left upper corner after zooming).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@4188 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c