benchmark-script: paths in set-files (includesfile) are interpreted as relative to the location of the set-file itself. comments allowed in set-files.
Example: 
If there is a program '/root/cpachecker/test/foo/prog.c'
and a set-file '/root/cpachecker/test/foo.set',

then the benchmark-xml-file remains unchanged (as is now): '<includesfile>test/foo.set</includesfile>'
however the set-file (foo.set) contains: ' foo/prog.c' (not 'test/foo/prog.c')

FALLBACK:
old set-files are handled as before, 
only a warning is printed before running the benchark.

AND:
comments allowed in set-files:
they start with '#' or '//' as first sign of the line and span over the whole line.
currently it is not possible to mix comments and files in one line.



git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@4362 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c