Fix important bug in ARTMergeJoin.
The covering relation was not updated when the resulting element of the
merged replaced the old element in the ART. This means that an element
covered by the old element was now suddenly uncovered, which should not
be the case.

This was never detected with predicate analysis, because there we have
two classes of AbstractElements: The elements of one class are
merged with other elements, but never cover; the elements of the other
class potentially cover, but are never merged.
with
other elements but will never

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@4906 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c