table-generator: no need to copy logfiles, we can access them with javascript/ajax. this should speed up the script.
option 'withoutlinks' removed, because there are no links.
the logfile-content is shown in the "chartWrapper"-div-block, that is now named "contentPane".

with local files this seems to work with firefox only.
chrome and opera do not allow access to local files through javascript.
with a webserver all browsers seem to work.
if the content cannot be loaded with ajax, a direct link to the file is shown.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@5361 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c