Change strategy of handling dead code.
There are two sets of states for the CFA of a given function:
- the set of states reachable from the entry node
- the set of states created during creation of the CFA

Previously, we wanted to defend against bugs in the cfa builder
and thus tried to ensure that these two sets are always equals.
This meant, during CFA creation we needed to detect all cases of dead
code, because otherwise the first set of nodes would have been smaller.

This strategy is quite difficult and has already caused some problems
for users (whenever they encounter a new type of dead code we had not
handled, there was an AssertionError).

Now in bug #71 there is the following code (simplified):

void foo() {
  return;
  while (1) {
    label: ;
  }
}

The whole loop is dead code, but only because the label is never
mentioned in a goto clause. This case of dead code is very difficult to
catch.

Additionally, bug #70 shows that there is a performance problem with our
detection of unreachable labels (the function mentioned in this bug
report is used for this).

Thus now I change the strategy and don't care anymore if the two sets
above differ. Instead after creating a function CFA, I use the set of
reachable states as the relevant set and delete all the other nodes.
Thus we don't need to handle all cases of dead code explicitly.

(fixes #71)
(see #70)

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@5668 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c