Catch OutOfMemoryErrors during loop detection.
Usually one should not catch Errors, but this is a special case.
The operation here is optional, and the analysis might well have the chance
to continue without it.
Furthermore, aborting the loop detection will usually free some amount
of memory immediately because a lot of objects fall out of scope,
so we probably don't hit the memory limit immediately again.

See #70 for a file which triggers an OOM during loop detection with
default heap size.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@5669 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c