CPAlien: Move return value processing to SMGState
Previously, the return value of the function was exlicitly handled by
tranfer relation, by creating and inserting additional stack objects
explicitly. Now, every CLangStackFrame has an object for its return
value created automatically, and SMGState has methods for obtaining the
object for the current frame, so that transfer relation can work with
it. This way, we can better protect the consistency of the underlying
SMG.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@7848 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c