benchmark-script changed, so that there is cloud-support for more tools. Now we support:
- Blast 2.7.1 (except some strange behaviour in case of Timeout. Then everything on CloudWorker is killed, we do not get any stdout-file (containing runtime, walltime, returnvalue, memoryusage) from the run, cgroup-bug?)
- ESBMC 2.21
- Threader (version from SV-COMP13)

To get this working, there is a small change in the CloudRunExecutor:
Instead of parsing a ugly escaped arg-string, we now parse a python-dump (string-representation of a dictionary).
This is not only much easier and readable (both in parser-code and in the cloudinput), 
it also allows us to send any data from the benchmark-script to CloudRunExecutor, without notifying or changing the VCloud.
Besides the original commandline for the tool, this data includes a flag for debug-mode and special environment-changes for tools. 

Other changes:
For a local run of the benchmark-script, the working-directory is set to the directory, the tool wants.
Now we log the toolname, not the python-object of the tool-class.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@8783 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c