CPAlien: Separate dropping frame from memory leak checks
Now we can decide with a configuration option whether we want to treat
memory that is not freed until the end of the main function as memleak.

This change was originally done on SV-COMP'14 parallel SMG effort,
cpalien2. It was originally authored by Philipp Wendler, later
integrated to cpalien trunk by Petr Muller.

Signed-off-by: Petr Muller <muller@redhat.com>

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@9438 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c