Copy everything from the branch SMG_Explicitly into the trunk. This includes: - The CPAs CPAlien and ExplicitCPA. - Some config files and test sets. - C programs for testing.
Because the code in the branch and the trunk were not yet successfully integrated,
but we want to have the code in the branch now,
I made copies of both CPAs and their config files instead of merging them into the existing ones.
This means, we now have (temporarily) two CPAlien and two ExplicitCPAs in the trunk.
All files in the packages of these CPAs were duplicated,
even those that were equal in both trunk and branch.
The CPAs from the trunk were untouched.
The CPAs from the branch SMG_Explicitly have gotten the suffix "2".
I added the suffix in the following places:
- at the end of the package name of the CPAs
- in the names of the config files
- in all places where these CPAs are referenced (e.g., in their config files)
I did *NOT* add the suffix anywhere else, including not in the following places:
- any class, type, method, or variable name
- the name of the configuration options
  (so both versions of the CPAs listen to the same configuration options!)

It is not possible to use both versions of ExplicitCPA or CPAlien
together in the same configuration,
you have to either use the one from the trunk or the one from the branch.
It is also not possible to mix between CPAs,
if you use both CPAlien and ExplicitCPA,
they have to be from the same source.

The rest of CPAchecker should work with both versions of the CPA,
I integrated them in all places where the compiler enforced it
(that's why there are a few changes in classes outside the CPAlien/ExplicitCPA packages).

The unit tests run through, but I did not ran any further test.

I added the mergeinfo property for SVN
so that SVN now believes this branch was cleanly integrated into the trunk
and won't propose any further merges.
The SVN history for all files copied was also kept intact.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/appengine@9685 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c