Replace the original CPAlien with SMG_Explicitly
The SMG_Explicitly was a branch where Alex D. and Stefan L. integrated
CPAlien and ExplicitCPA. During the SV-COMP14 competition efforts, the
code was put to the trunk as a copy to allow participation. This commit
integrates the copy into the original thing. The partial commit history
of the changes is lost.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/appengine@9738 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c