Deprecate the use of java.nio.file.* by replacing it with a wrapper class.
To run the CPAChecker on the Google App Engine it must comply with Google's
whitelist. Therefore every use of classes in the package java.nio.file.*
must be deprecated. To replace the affected classes a new wrapper class was
added to the common library. The wrapper org.sosy_lab.common.Path should be used
instead of the aforementioned classes.


git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/gae-port@9967 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c