We have a Sylvester release again! Release 1.3 tags the development progress of the last year and contains several new interesting features that were introduced in 2013.
2013 was again a very good year for the CPAchecker project, in particular,
we can be proud of the achievements in SV-COMP 2014.
Thank you for your contributions as developers,
and for your patience and suggestions as users.

Happy New Year!
I wish you all the best for 2014, especially for your CPAchecker-based verification projects!
Let's continue to intensively collaborate on this large and successful project.

Yours,
Dirk



git-svn-id: https://svn.sosy-lab.org/software/cpachecker/tags/cpachecker-1.3@10215 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c