Extract CPU time, wall time and other statistics about a run from AppEngine log.
When a job has been run a log entry concerning this particular request has been
created. When the job is being run its unique request ID will be saved alongside
so the according log entry can later be identified.
When the job has finished running the following statistics will be retrieved
from the associated log entry upon the first request of the job resource:
- Estimated cost in US Dollar
- The host that ran the job
- The latency (wall time)
- The start time
- The end time
- The pending time (time job spent in the queue)
- CPU Mega cycles. Can be converted to aproximated seconds.

Sometimes a job crashes and never transitions into a finished state. However the
request still ends at some point. This is reflected in the log and thus the
status of a stuck job can be corrected afterwards.


git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@10565 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c