Update SosyLab common to 0.396. Introduces a factory for configuration builder.
The need for a factory to create ConfigurationBuilder instances derives from a
problem with CPAchecker on Google App Engine. It is possible that classes create
their own Configuration instances and therefore overwrite options. However, on
App Engine a certain set of default options needs to be set at all times and
must not be overwritten. Hence, using a factory to create ConfigurationBuilder
instances makes it possible to provide a builder that will ensure this constraint.


git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@10831 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c