bug fix: due to the bug fix/change in commit 11244, it now could happen, that BREAK is signaled for the first successor of a parent with two successors, removing the second successor from the ARG.
If, however, the first successor is covered, the BREAK signal is ignored, and the analysis proceeds with the second successor. This state is already destroyed (removed from ARG), leading again to an inconsistent analysis.

This problem only arises with the delayed-BREAK semantics of the ExplicitGlobalRefiner. Thus, an additional check, whether or not the current state is a target state avoids this problem, and BREAK is signaled despite coverage if the current state is NOT a target state.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@11251 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c