Refined the strategy to determine where to restart the analysis after a successful refinement. Before it was either TOP or BOTTOM encoded in a boolean flag/option named "doLazyAbstraction".
Semantically, the COMMON strategy is new, which means that the analysis is restarted from the lowest ancestor common to all nodes, if more than two refinement roots where identified

Also renamed the option to "restart", analogous to a similar option in predicate analysis, aslo renamed the prefix to refinement (instead of refiner) for the very same reason.

For the standard (non-global refinement) nothing is changed so far, but will be as soon as the two strategies (global and non-global) are unified.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@11492 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c