Start extracting program state from SMGState
SMGState is a superpowered class now: I'm working towards making it
responsible for handling the CPAChecker notion of state, and making
CLangSMG responsible for keeping the program state. SMGState now allows
to pass the read-only copy of the SMG out of the class for reading. In
the future, I would like to have SMGState holding a non-modifiable
CLangSMG, and only constructing new states after doing all the
modifications externally. This should prevent us from all the bookeeping
about which SMS should and should not be modified during execution.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@11532 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c