Revert "CPAlien: Put stack-based modification behind SMGStateBuilder"
This reverts commit 11595.

This and the following 25 commits revert most of the changes to the SMGCPA
since r11178 (2014-02-17).
These commits have introduced a lot of regressions
(wrong answers and exceptions).
The plan is to revert these commits, fix the remaining problems,
and then work forward again, this time without introducing all these
regressions.

Commits that were found out to be irrelevant for the regressions
were not reverted as far as possible.
Nothing outside the SMGCPA package is affected by the reverts.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@12138 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c