HashMap and HashSet have a constructor that takes one int value as "capacity". This constructor should in general not be used, as it does not do what most people expect. There are the more appropriate Guava methods Maps.newHashMapWithExpectedSize() and Sets.newHashSetWithExpectedSize() which should instead be used.
What is usually desired when using these constructors/methods is
to have a map/set where so many elements can be put in
without any resizing.
However, both HashMap and HashSet already resize
before the capacity is reached.
In order to avoid resizing in cases where the final size is known,
a larger value needs to be passed as capacity,
which is what the Guava methods do.

For maps/sets which are expected to be empty,
using the constructor with capacity 0 is ok.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@13233 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c