The benchmark script truncates log files that are too large (> 20 MB). This was not working reliably and is now fixed.
The problem was that the whole file was read into memory
before attempting its truncation.
This failed on very large log files (with several GB) due to an OutOfMemoryError,
with the result that the file was kept unchanged.
This led to problems with the VerifierCloud,
which then had to handle the large file.

Now the file is truncated without being read into memory.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@13318 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c