Mark almost all configuration options as "secure". In the future, all configuration options that do not have secure=true will be forbidden in some modes of operation of CPAchecker (primarily when executing CPAchecker via the web frontend of the VerifierCloud).
Please also set secure=true for all new configuration options,
except if an attacker could exploit this option to gain control
over the machine, for example by executing arbitary code.
This is mostly the case if the option specifies the name of a tool
that will be executed by CPAchecker.
One example we have for this is the option "parser.preprocessor",
which is not secure.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@14001 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c