bug fix: for some reasons, when combining the value analysis and the predicate analysis, both with refinement
It happened that the VA refined a path (A), then the predicate analysis refined one or more path, and then, path A had to be refined again, triggering an exception in the value refiner (repeated counterexample). This now only is thrown, when a path is found in two refinements that happen *directly* after each other. However, it is still strange that *infeasible* paths are found multiple times, which remains unsolved for now

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@14993 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c