When running value analysis (VA) and predicate analysis (PA), merging/covering and refinements of both CPAs could lead to the state, where in two subsequent refinements, two identical error paths were found, through different parts of the ARG, triggering an repeated-cex-exception.
So now, when running VA+PA,
the refinement root is set to the lowest common ancestor
of those states that are covered by the states in the subtree
of the original refinement root.

This is only done when running both the VA and PA in parallel.

If either of the two is run in isolation,
this heuristic is neither needed
nor did it result in an improvement,
but a slight degradation,
so it is only active when both VA+PA
run in parallel.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@15112 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c