Updates Z3 version to the top of the OPT branch.
Updating the interpolation API was very tricky.
There's a bug in the python code which generates
the Java binding: for any Z3 code requesting
pointers, the API writes out the pointer-to-pointer value, instead of writing
the pointer value.

Z3InterpolatingProver#getSeqInterpolants wasn't updated as it isn't the default option, left for TODO.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@15128 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c