Cast symbolic values and constraint expressions correctly.
Cast symbolic values in assignments and expressions to the required type.
This creates a new SymbolicValue object with all the same properties but the type.

Handle casting of constraint expressions and symbolic values in creation of BitvectorFormulas.
Always cast operands' formulas to the given expression's calculation type and the expression's complete formula to the expression type.



git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/symbolic@15248 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c