Python renaming part I (see #199)
This and the following commits rename methods, variables and fields
in most of our Python scripts such that they conform to the
standard Python naming scheme: lowercase_with_underscores (see PEP 8).
Currently the naming is a mix between this and javaMixedCaseStyle.

The renaming is primarily applied to methods and fields,
with variables being renaming mostly if they happen to have a similar
name (renaming all variables consistently would be too much effort).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@15650 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c