Refactor value analysis refinement to become more generic.
Use dependency injection in most refiners that are somehow related to ValueAnalysisRefiner to allow the use of an arbitrary feasibility checker and strongest post-operator and to use these permanently.
Previously, ValueAnalysisFeasibilityChecker was created at different locations independently.
In addition, a ValueAnalysisTransferRelation was created at different locations and used as strongest post-operator.

The newly introduced StrongestPostOperator allows the use of different transfer relations through one unifying interface. However, many classes and methods of value analysis refinement still rely heavily on the strongest post-operator to return a ValueAnalysisState.
ValueAnalysisRefiner also relies on the use of a ValueAnalysisFeasibilityChecker for evaluation of ARGPaths.
It should be evaluated how or if these issues should be resolved.

Currently, strongest post-operator and feasibility checker are commonly generated in the refiner's create(CPA) method and injected by using the constructor.



git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/symbolic-cegar@16330 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c