substitute implementation for Mathsat
This is rather questionable, so I have not added it to
AbstractUnsafeFormula manager.
The problem is comparing the different ASTs, which is done here using
to-string conversion.
NOTE: the performance of this method is quadratic in size of the formula, as each item has to be converted to a string first.
However, policy iteration needs this for MathSAT integration.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@16436 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c