reverting r16176 and r16196 + fix for Value-RTT-strenghening.
Dropping the frame has to be done with the assignment of the return-
variable. This was changed in r16176 to get some better results for RTT-
and Java-related analyses. The current commit reverts the change and
solves the corresponding problem with the RTT-CPA by using the oldState
that is available for strenghening.

Renaming variable 'state' in ExprValueVisitor, because there is an
equal-named field in ValueTransfer, which was used in a inherited
visitor.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@16537 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c