Provide different possible less-or-equal operators for ConstraintsCPA.
Move them (and the ConstraintsState) to their own package [..].cpa.constraints.domain
for a clearer structure.

ImplicationLessOrEqualOperator checks whether a state's constraints are implied by
another state's constraints.
SubsetLessOrEqualOperator checks whether a state is a subset of another state.
AliasedSubsetLessOrEqualOperator checks whether a state is a subset of another state,
if all symbolic identifiers of the first state have an unambigious alias in the other.

To allow these different operators, provide an explicit AbstractDomain.
As we need a solver in ImplicationLessOrEqualOperator, move the solver creation from the
transfer relation to the CPA so we only have to create a solver once in the cpa.



git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/symbolic-cegar@16624 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c