New prefix preference for error path slicing trying to evade floats and binary operations.
The new prefix preference is based on the domain type score of variables, but includes a new
set of variable classification, IntArith.
IntArith contains all variables that are not floats and that occur only in simple operations that are not
binary or, binary and, or binary xor.

Score computation for this prefix preference is based on multiplication and
assigns IntAdd (and unclassified variables) a high malus/factor to avoid floats and binary operations.



git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/symbolic-cegar@16650 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c