Revert changes to SymbolicValue.equals(..) in r16671 and introduce new method for comparing symbolic values for their symbolic meaning.
The represented memory location has to be considered in symbolic values'
equals and hashCode methods to ensure ConstraintsPrecision.isTracked(..) works
as intended. Since it compares symbolic values based on the C code/memory
locations they represent, it must be possible to add two constraints with the
same symbolic meaning, but different represented memory locations to the same
constraints state.
This is necessary because it is not possible to reliably determine the symbolic
meaning of a specific code-representing constraint at the time of refinement
without using the existing precision.
(e.g. a < 5 can be s1 < 5 in refinement but s1+10 < 5 in the next iteration of
CPA algorithm)



git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/symbolic-cegar@16677 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c