Fix bug in SymbolicValues.representSameCCodeExpression(..), add unit test for it.
Fixes a bug that assumed that two symbolic expressions of different types
represented the same C expression if they didn't represent a memory location
at all.

Example: a > 0, !(a > 0) result in constraints s1[a] > 0, !(s1[a] > 0).
the constraints themselves do not represent a memory location and are of
different types (less than vs. logical not), so SymbolicValues
returned them to be equal in respect to the C code they represent.


git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/symbolic-cegar@16697 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c