ConstraintsCPA: Implication-leq-operator considers definite assignments of state.
Symbolic identifiers with known definite assignments are not immediately replaced
with their concrete values in the strengthening of Value Analysis, but only in
the next iteration. This is because CompositeCPA's strengthen mechanism works
on the results of the transfer relations and does not consider the
already existing strengthen results.
Because of this, the implication operator has to consider definite assignments,
as Value Analysis does not yet reflect this information.



git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/symbolic-cegar@16786 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c