Removes SSA tracking from the counterexample.Model
This is a large change with a potential to break things.
SSA handling in the model mixes two abstraction layers, is inconsistent (indexes are stripped out from Variables, but not from Functions) and is somewhat confusing.
Moreover, it makes the SMT solvers code (who create the Model object) tightly coupled to CPAchecker's logic.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@16931 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c