Change implementation of UnsafeFormulaManager.replaceArgsAndName() to replace both arguments and name at the same time instead of in two steps. Also remove replaceName() because it is used only in one place and can be replaced by replaceArgsAndName() trivially. This new one-step method may be slightly more efficient.
It also has a further side effect.
Because SmtInterpol does not support global declarations
(i.e., declarations that live even after the stack is pop'ed)
we sometimes end up with terms that are not defined in SmtInterpol
anymore. If we call replaceArgsAndName() on them to instantiate them,
the previous code would have tried to create first a term with the new
args and the old name, which would fail due to an undefined term.
The new code works in this situation if the new name happens to be
defined.

This is an improvement in rare situations happening for example
with sliced path prefixes, where the code to create them calls
uninstantiate while the solver stack is non-empty
(and uninstantiate may happen to define terms).
It is, however, not a full solution, which would be to never define
terms (and thus not call (un)instantiate) with a non-empty stack,
or wait until SmtInterpol has support for global declarations.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@17092 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c