VA: Add option to control whether to replace symbolic values in AssigningValueVisitor.
When SymbolicDelegatingRefiner is used in symbolic execution refinement,
value analysis _must_ replace symbolic values in unambigious assumptions
so that the semantics of the explicit strongest-post operator and the
value analysis transfer relation are the same.
This removes the possibility to create a relation for symbolic identifiers, though.

Example: VA replaces symbolic values.
For a = s1 + 1 with s1 symbolic identifier and assumption (a == 1) the information
that s1 = 0 is lost, but only a = 1 is stored. In case another variable b = s1,
we don't get to know that b = 0.



git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/symbolic-cegar@17128 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c