ConstraintsCPA: Concrete values in constraints derived of memory locations  store the represented memory location.
If a constraint is created using a memory location whose value is a concrete value, store
the memory location in the used concrete symbolic expression created for the
constraint.

Fixes a bug where a refined ConstraintsPrecision failed to prove the counterexample as
infeasible since the value of a program variable was symbolic during refinement, but
replaced with a concrete value in casual analysis due to the greater precision (from
previous refinements). The resulting constraint then was not covered by the precision,
as the concrete value did not represent any memory location, falsely.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/symbolic-cegar@17164 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c