VA: Do not replace symbolic values in assumptions with a single possible value, by default.
Replacing symbolic values in value analysis discards information about the symbolic identifiers
contained in it. The ConstraintsCPA computes the definite assignment and will tell it to the
value analysis in the next transfer's strengthening.

Currently, the value analysis only replaces assigned symbolic identifiers, though, and does not
resolve symbolic expressions.



git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/symbolic-cegar@17194 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c