bug fix / improvement: when extracting a precision for the tree of interpolants, do not collect interpolants beyond a state that has the interpolant FALSE.
As the tree is build in a lazy fashion, it can happen that a non-trivial(!) interpolant is available beneath a FALSE interpolant. This can happen for example if a later refinement finds a contradiction (interpolant FALSE) higher up in the tree than a former one (having (non-)trivial interpolants there and beneath).

However, these interpolants are not needed to excĺude the former error, as they are subsumed by that latter, so just disregard them.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@17240 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c