Z3-wrapper: adding option to disable DIV and MOD for non-numerals.
The operations DIV and MOD are expensive (non-linear arithmetic).
Z3 allows to use them for non-numeral operands,
but perhaps this leads to timeout.
This option allows to use UFs for all cases,
where the second operand is no numeral.

Additionally the code for DIV and MOD was moved out of the super-class,
because we do not want it for Rational-theory.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@17968 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c