BAM: adding a second search for a path.
We have refiners that search for an alternative path
after repeated counterexample. BAM depends on finding
the cut-state for the refinement in the original path.
This commit inserts a second search for such a path,
right before the cut-state is used to cut the BAM-ARG.

TODO:
This commit might fix some problems, but it is only a
temporary solution. We should remove the dependency
from a special path.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/DelegatingBAMRefinerLDV@18538 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c