Adds new feature implementation: CFASequencer.
This commit contains the implemented feature of cfa sequencing. 
This includes new files to postprocess the cfa 
  (see package org.sosy_lab.cpachecker.cfa.sequencer),
new cpas to ensure a correct thread scheduling  
  (see package org.sosy_lab.cpachecker.cpa.context),
a MultiCallstackCPA to handle multiple callstacks (for every thread one) 
which appear in a concurrent program 
  (see org.sosy_lab.cpachecker.cpa.callstack).

This commit contains also some changes on the cfa model such as new edges and 
adaptions in the checkers model to work with them.

For detailed information run
  "svn diff -r 16821:<revision number of this commit>"  
or checkout the git repository 
   https://Tantalos@bitbucket.org/Tantalos/sequencer.git



git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/cfaSequencer@18819 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c