Update to newest version of SoSy-Lab Common. This includes a split of ShutdownNotifier into two classes, one for issuing shutdown requests and one for receiving them. Most code does not need to change because it only receives shutdown requests, but all code that calls requestShutdown() needs to have a ShutdownManager now.
This prevents that a sub-component that should receive shutdown
requests accidentally issues a shutdown request that affects its parent
component as well.

The only behavioral change of this commit is related to this problem:
BMCAlgorithm and all invariant generators called requestShutdown()
on the ShutdownNotifier that was given to them,
i.e., they could unexpectedly shutdown the component that instantiated
them.
This commit now gives the BMCAlgorithm and the invariant generators
a child ShutdownManager instead, such that they can issue shutdown
requests internally and for example shutdown the CPA and the solver
that they use, but do not affect other surrounding components.

All other changes only pass around ShutdownManager instances
instead of ShutdownNotifier instances wherever the former is needed.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@18901 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c