Add classes for left hand side expression handling.
The bring the heap array approach to a further step, we need to change
the handling of C left hand side expressions in order to use the SMT
arrays more efficient.
This simply copies the code of the original classes
LvalueToPointerTargetPatternVisitor and PointerTargetPattern from the
pointeraliasing package. No changes were made in terms of functionality.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/heaparray-integration@18960 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c