Move preconditions project to orphaned projects.
The code is unmaintained and misses tests
(most of the currently existing tests are deactivated).
There are two major upcoming refactorings where this code would need to
be adjusted:
- JavaSMT traversal API
- ARGPath changes regarding the removal of MultiEdges
Without regression tests and a maintainer these refactorings
would be hard to do and probably introduce errors,
thus the code is removed from trunk and put into /orphaned/precondition
(with next commit).

This affects the seghirPrecondition* configurations
and the option cpa.predicate.abstraction.elimDeadVariablePreds,
which depends on the same code,
as well as the code for SMT AST matching.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@19083 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c