Added mutex support
Added a stubbing for mutex_init mutex_lock and mutex_unlock.
Verification task succeeded with atomic mutex program in sv-compilation.

The stubbing is implemented in a embedding of the pthread functions.
With this approach the variables used to control the scheduling will be cept to a minimum.

No check for illegal use of pthread functions such at
  - trying to lock a destroied mutex
  - trying to unlock a locked mutex by an other thread



git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/cfaSequencer@19138 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c