Added methods to MutableCFA - fixed bug in function cloning
-added assertions which affirm the correctness of the new crated cfa.
-every used function in cfa is a cloned function now
-added function remove, setMainFunction to MutableCFA


All functions will be cloned
for so that every thread has it's unique function. The function
Calls was adapted to this new functions
Original functions were reused if their weren't used by a thread
already.
This results in an error because the changed function calls wasn't
tracked as used function but as new functions. This results in a 
doulbe use of functions.
 




git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/cfaSequencer@19190 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c