Remove code for adding retention constraints
The code was commented out after a fix in r19578. That change was
necessary, as the previous versions contained a bug with incrementing
the SSA index when storing a new value in the SMT array.

We now remove the complete code that did add the retention constraints
as it doesn't seem to be necessary any more.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/heaparray-integration@19597 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c