Implement preliminary quantification of initializer
When having a C initializer, e.g. for an array, we can now use a forall
quantifier to push this into SMT formulae, given the SMT solver is
capable of quantifiers and arrays.

ATTENTION: By now, the name for the array formula is hard coded, because
there is an issue with getting the proper name (cf.
AssignmentHandler.java:287). We need to fix this before doing real
testing—right now, the implementation is only tested with two small C
files.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/heaparray-integration@19731 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c