Add basic code to hand over AutomatonASTComparators
In some cases, variables will not be classified relevant by the
VariableClassification but are relevant because we have an assumption on
them in an automaton.

The idea of this approach is to parse the given automata and get their
AutomatonASTComparators before CFA creation; they will be given to
VariableClassification in order to find variables that are only relevant
for the analysis because of a specification automaton, i.e., a variable
that would be omitted without the specification.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/muauto@20084 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c