Passing the updated state to child CPAs in CompositePrecisionAdjustment
I know this is very hacky, but currently there is _no_ other way to
communicate between the two CPAs if they both perform abstraction in
"precision adjustment".

The alternative would be to perform abstraction in transferFunction, but
it does not have access to the ARGState or the ReachedSet, which I use
to find predecessors.
I'm open to negotiations for a more acceptable API.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@20166 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c