Change PredicateStaticRefiner into a wrapping ARGBasedRefiner.
Previously it was integrated into PredicateAbstractionRefinementStrategy.
This lead to some irrelevant work being done for static refinements
(such as performing refinement selection).
It was also difficult to decide whether a repeated counterexample
was because of previous static or interpolation-based refinement.
Furthermore, one could enable static refinements for Impact
configurations, but it would silently stay unused,
or it could get enabled for PredicateExtractionRefiner
where it is not desired.

Now it is a refiner that wraps the actual refiner
and on first call does static refinements and afterwards
always forwards all calls to the actual refiner.
The actual refiner is not even called for the static refinement,
making it much easier to understand which code is used in which
situation, and avoiding the need for special-casing static refinement
in the standard refiner.
This also makes it possible to combine static refinements with
ImpactRefiner, which was not possible before.

Because of this, the decision whether heuristically-gathered predicates
are split into atoms or not is now controlled by a separate
configuration option and not the same one as for interpolants
(just like there is a separate option for splitting invariants).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@20294 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c