Allow the combination of Impact refinements and predicate abstraction.
Typically, with Impact refinements, the precision will always stay empty
and thus every abstraction "computation" during the analysis will simply
return "true".
However, we could have predicates in the precision coming from other
sources (e.g., predicates initially read from a file, or from static
refinement), and they are used in the normal way during abstractions.
This can be seen as "seeding" the Impact analysis with stronger abstract states.
The biggest advantage for now is probably to make the static refiner
work the same way in combination with the Impact refiner
as it does with the predicate-based refiner.

Before this change, this combination could already have been configured,
but the outcome from the abstractions would not really strengthen the
abstract states because the abstraction formula would consist only of
combinations of some fresh boolean variables without connection to the
program variables. Now we actually add predicates over the program
variables to the abstraction formula, just like Impact refiner does,
which means the output of the abstraction has exactly the same effect
as if it would have been added to the abstract state by an Impact refinement.

The implementation of this change violates the encapsulation of
SymbolicRegionManager, like the existing code already does anyway.
Getting this implemented nicely is tricky because SymbolicRegionManager
has no way of determining what is supposed to be a predicate
(predicates do not need to be SMT atoms, otherwise it would be easy).
Maybe another solution has to be found, e.g., like lifting
the SymbolicRegionManager to a SymbolicAbstractionManager.
The justification for adding this hack now is to separate
the actual feature (the code for which is quite small)
from a potentially big refactoring to get it implemented nicely.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@20323 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c