Set default of cpa.predicate.targetStateSatCheck to false as a preparation for removing this option altogether. The PredicateCPA now does a SAT check at target states in precision adjustment if the precision says so, so there is no need to have the explicit SAT check in the transfer relation as well.
This is closer to the principle of lazy abstraction and more flexible,
because we now do the SAT checks only if the precision tells us, and not always.
The eagerness of having the SAT check always only pays off
if there are a lot of cases where the last block before the target state
is contradicting in itself (so we can rule out the target state just by
looking at the last block without the need for a refinement).
However, this is expected to be rare.
Furthermore, if we once find it to be worth the cost in general,
we could still simply seed the initial precision appropriately.
The increased flexibility of being able to choose when to do such SAT
checks is beneficial for example in configurations that combine the
PredicateCPA with the ValueCPA. There we usually would like to use SAT
checks only when the ValueCPA cannot handle a path, but previously we
had to do these target state SAT checks always because we could not have
them on some paths and not on other paths.

Another simplification is that now we can control whether SAT checks
are always done for blocks with the single option
cpa.predicate.checkBlockFeasibility, previously we had to use
cpa.predicate.targetStateSatCheck for blocks ending in a target state
and cpa.predicate.checkBlockFeasibility for all other blocks.

Some configurations and components which relied on the
targetStateSatCheck needed to be changed in order to work correctly
no matter whether the targetStateSatCheck is true or false.
For all cases that I know this was done in the previous few commits.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-no-targetStateSatCheck@20460 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c