Add expression type fixing for automaton assumptions
When using a type or variable in an automaton that is declared in a C
header file, the type information cannot be inferred during automaton
parsing because the information is not yet present there. We introduce a
method to fix this afterwards, when all information is present.

ATTENTION: This is not yet finished an can cause problems!

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/muauto@20819 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c