Change slightly how PredicatePrecision works.
Originally, when doing a lookup it would look into the multimap for
function-specific predicates, and if it finds something there, return it,
otherwise return the set of global predicates.
The various methods for adding predicates to a precision made sure
that global predicates where also added to all functions for which
specific predicates are present, such that the lookup result always
contains the global predicates without doing a union of sets during
lookup. The constructor of PredicatePrecision, however, did not do
this and allowed to create precisions that have a given predicate as a
global predicate, but not in all functions.
Then r18454 changed this behavior and added a union computation of all
sets during lookup such that now a lookup with getPredicates() was
guaranteed to contain the global predicates, no matter how the precision
was created. The individual sets where not updated, however, such that a
lookup with getPredicates() could give different results than
getFunctionPredicates().get().
This was also rather inefficient, because most code was still doing the
eager union computation explained above while getPredicates() did an
additional lazy union computation.

(Everything written above applies not only to the function-specific
predicates, but also to the location-specific predicates and the
location-instance-specific predicates.)

This commit now changes the constructor of PredicatePrecision such that
it always computes the union of all predicate sets eagerly,
i.e., what getPredicates() did lazily since r18454.
This keeps the same behavior of getPredicates() as it had since r18454
(except possibly for iteration order of predicates),
but might change the result of getFunctionPredicates() etc.
to make it consistent with getPredicates().

At the same time, this commit makes the various creation methods more
efficient (always two copies of the input map now, instead of up to three)
and ensures the same iteration order in all predicate sets.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@21004 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c