Extend automaton parser
We add the keywords "TRACK", "UNTRACK", and "ISEMPTY". Their syntax is
> MATCH {$1=malloc($?)} -> TRACK("<set>", $1) ...;
> MATCH {free($1)} -> UNTRACK("<set>", $1) ...;
> MATCH EXIT -> ISEMPTY("<set>");
where "<set>" is the name of a set; this can be an arbitrary name to
identify the tracked variables in the set.

The idea is to provide a tracking, e.g., for pointers that get
initialized by a malloc call, in order to see whether the pointer is
freed correctly.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/muauto@21367 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c