Add class that encapsulates the specification of a CPAchecker run.
The information in this class is not intended to be actually used by
components of CPAchecker, instances of this class are only meant to be
passed on to sub-analyses, such as invariant generators, counterexample
checks, etc., all of which typically need to use the same specification
like the main analysis.
Previously we copied the value of the "specification" configuration
option to the configuration of sub-analyses, which is ugly, error-prone,
and lead to unnecessary repeated work (parsing of automata).
This is now not necessary anymore, we can simply pass on a Specification
instance, which can be used in the sub-analyses when creating CPAs.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@21498 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c